auth-server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
bcrypt==4.0.1
pytest==7.4.3

# Session support
flask-session==0.5.0

# Google OAuth
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1